Fixes #1986 - Finishes migration to Intern 4.
- Adds 'browsers' optional argument when running tests
- Changes folder structure separating functional and unit tests completely
- Also fixes #1977 and #1981.
