+++ tests/.eslintrc
@@ -1,5 +1,9 @@
 {
   "globals": {
-    "intern": true
+    "intern": true,
+    "process": true
+  },
+  "rules": {
+    "no-console": "off"
   }
 }
