+++ tests/functional/lib/window-helpers.js
